Fix offset issue (#234)
* fix reading

* better example

* fix offset

* Fix example

* fix bug

* fix one msg

* fix name conflict

* fix unit test

+++ onnx_extended/ext_test_case.py
@@ -392,7 +392,7 @@ class ExtTestCase(unittest.TestCase):
         if folder and not os.path.exists(folder):
             os.mkdir(folder)
         res = os.path.join(folder, name)
-        if os.path.exists(res):
+        if clean and os.path.exists(res):
             os.remove(res)
         return res
